add configurables

+++ src/main/java/dev/xkmc/l2transport/content/upgrades/ThroughputUpgrade.java
@@ -13,7 +13,7 @@ public class ThroughputUpgrade extends Upgrade {
 	}
 
 	@Override
-	public int getMaxTransfer(int cd) {
+	public long getMaxTransfer(long cd) {
 		return cd * throughput;
 	}
